+++ workspaces/core/src/avatar/private/types/AvatarProps.ts
@@ -9,6 +9,7 @@ type AvatarProps = {
   size?: SizeProps;
   statusOrigin?: StatusOriginProps;
   isStatusActive?: boolean;
+  isRounded?: boolean;
 };
 
 export default AvatarProps;