+++ workspaces/core/src/label/Label.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import classNames from '../private/utils/classnames';
 import LabelContainer from './styles/LabelContainer';
 import type LabelProps from './private/types/LabelProps';
 
@@ -9,7 +8,7 @@ const Label: React.FC<LabelProps> = ({
   backgroundColor = 'bg-bulma',
   iconLeft,
   iconRight,
-  size,
+  size = 'xs',
   isUppercase = true,
 }) => (
   <LabelContainer
@@ -20,27 +19,9 @@ const Label: React.FC<LabelProps> = ({
     iconLeft={iconLeft}
     iconRight={iconRight}
   >
-    {iconLeft && (
-      <span
-        className={classNames(
-          'me-1',
-          size === '2xs' ? 'text-[0.75rem]' : 'text-[1rem]'
-        )}
-      >
-        {iconLeft}
-      </span>
-    )}
+    {iconLeft && iconLeft}
     {children}
-    {iconRight && (
-      <span
-        className={classNames(
-          'ms-1',
-          size === '2xs' ? 'text-[0.75rem]' : 'text-[1rem]'
-        )}
-      >
-        {iconRight}
-      </span>
-    )}
+    {iconRight && iconRight}
   </LabelContainer>
 );